+++ src/views.rs
@@ -1,5 +1,5 @@
 // track - A quick-and-dirty CLI tool for tracking parcels.
-// Copyright (C) 2020, 2022 Jean-Philippe Cugnet <jean-philippe@cugnet.eu>
+// Copyright (C) 2022 Jean-Philippe Cugnet <jean-philippe@cugnet.eu>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -13,21 +13,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-//! A quick-and-dirty CLI tool for tracking parcels.
+pub mod tracking_info;
 
-#![warn(rust_2018_idioms)]
-#![warn(clippy::redundant_pub_crate)]
-#![warn(clippy::unwrap_used)]
-#![warn(clippy::use_self)]
-#![deny(missing_docs)]
-#![deny(unused_must_use)]
-#![forbid(unsafe_code)]
-
-use eyre::Result;
-
-use track::Track;
-
-fn main() -> Result<()> {
-    color_eyre::install()?;
-    Track::run()
-}
+mod askama_filters;